Add lab tests
This patch adds another test suite, that will run a suite of tests that
interact will (real) network devices, checking for the desired
operational state of the network device by inspecting it's
running-config through a vty session. These tests are designed to be run
as a Job in a Kubernetes cluster where they interact with virtual
Devices that are run using Containerlab.