cisco-nxos-provider: DeletingUpdate implementation
The gnmiext client does not provide a way to delete objects from the
data tree. Until now we worked with static objects. This is now
different with, e.g., VRFs, which can be arbitrarely added and removed.

DeviceConf methods like Reset() can now return the DeletingUpdate
implementation of the Update interface. If a call to  gnmiext.Update
happens and the update objects are of this type then the gnmiext.Update
will issue a gNMI request with mode `delete` to remove the objects
contained in the data tree specified by the xpath.