cisco-nxos-provider: decouple OSPF from interface
* We are using the pkg iface to configure several options related to
interfaces, including L3 settings like routing processes attached to the
interface. While this functionality has not been implemented for OSPF
yet, we have decided to decouple this relationship and move everything
related to routing processes to their respective routing packages (e.g.,
ospf in this case).

Consequently, in this commit we add the configuration of the OSPF
settings of an interface to the ospf package to enable the equivalent to
these CLI commands on a cisco nexus 9k:
```
interface eth1/1
ip ospf network point-to-point
no ip ospf passive-interface
ip router ospf UNDERLAY area 0.0.0.0
```

The current implementation assumes that the interfaces are properly
configured (i.e., they are "routed" interfaces)