cisco-nxos-provider: add port-channels to pkg `iface`
To implement vPC feature for this provider we require configuring
port-channels first. With this commit we enable configuring the
equivalent to:
```
interface port-channel10
  description ThisIsADescription
  switchport
  switchport mode trunk
  spanning-tree port type network
```

where the description, the switchport mode and spanning tree are
configurable via Options. We include the possibility to add physical
interfaces to the portchannel by means of the `WithPhysicalInterface`
Option. Using this option is equivalent to something like:
```
interface ethernet1/11
  channel-group 10 mode active
```