cisco-nxos-provider: add Exists method to gnmiext
In some cases it is necessary to check if an xpath exists. Currently we
rely on the `Get` method for that, which has to Unmarshall the json
data. We introduce the `Exists` method to avoid this operation.
Likewise, we change the interface getter method from pkg `iface` to use
them.