cisco-nxos-provider: handle existing trustpoint
Trying to replace the configuration of an existing trustpoint will fail
with "disassociating rsa key-pair not allowed when identity certificate exists".

In order to avoid this, we check if the trustpoint already exists and return
an error if it does. This can be checked in the caller and handled appropriately.

Additionally, when deleting a trustpoint, it now also delete the associated
keyring that is not automatically garbage collected by NX-OS.