cisco-nxos-provider: fix `pkg` ISIS
* We were previously adding an interface to an ISIS process by updating
  the yang data model at two different locations:
  - System/isis-items/if-items/InternalIf-list[id=eth1/1] (ReplacingUpdate)
  - System/isis-items/inst-items/Inst-list[name=UNDERLAY] (ReplacingUpdate)
These two updates caused the interface to flap (probably as they
reference each other and a change in one triggers an update in the
other).  This behavior can be avoided by adding the interfaces directly
into the second path and performing a single ReplacingUpdate. As a
consequence, we change the data model for ISIS and include a list of
member interfaces as a field.
* Forbid configuring ISIS on port-channels
* Add support to enable `bfd` on the interface by adding an Option to
  the Interface constructor