Auto-detect host.docker.internal ip address as device sample target
During testing it's often useful to port-forward a real or virtualized
device (e.g. via Containerlab) to the localhost in order to use the
device as a test case. For this purpose this patch automatically reads
the ip address of `host.docker.internal` if set and applies it to the
`Device` sample resource when using the Tilt setup.