cisco-nxos-provider: check if stp entry exists for physif reset
While trying to replace the yang subtree for the spanning tree
configuration, the operation fails with an error indicating that the
interface is not in Layer2, if the spanning tree subtree has not been
present before. To workaround that situation, we now first check if the
spanning tree yang subtree exists, before trying to replace it.