cisco-nxos-provider: build physif object from remote
* We want to avoid pushing new updates to the switch if its config has not
changed. This happens quite often, everytime a new reconcile is
triggered. Thus, we introduce methods for a user to retrieve the state
from the remote switch and populate our structs.
* We add a `FromYGOT` method to the DeviceConf interface. This method
  takes a gnmiext.Client as parameter. The `FromYGOT` function uses this
client to request various parts of the yang tree and populate the
DeviceConf implementation directly.
* Notice that an object is represented by multiple xpaths and the values
  therein, and that these xpaths often include lists with key values in
the middle of the path. For example a VRF name is often inserted in the
middle of an XPath. As such, it is not trivial create a list of Updates
to be requested with the client. Instead, we need to issue requests that
depend on the values returned by previous requests.
* We also add an `Equals` method to the DeviceConf interface. This
  method implements a comparison that can be utilized within a reconcile
iteration to check the differences between two objects. For example:
```
// declare an object and pre-set relevant key values, i.e., interface name
rConfig, _ := iface.NewPhysicalInterface("Ethernet1/1")
... // do error handling

// update rConfig with data from remote switch
err = c.Pull(context.Background(), rConfig)
... // add err handling

// lConfig is the object populated with CRD values
areEqual, err = rConfig.Equals(lConfig)

// reconcile or skip
```
* We implement an example for `PhysIf` and leave all other methods
  unimplemented for future work.
* Refactor pkg `Features`. Features is now a struct that incudes a List
  of features (a slice of strings). This change is done for convenience
after the introduction of the Equals method.
* Addressed several linting and static check errors