cisco-nxos-provider: allow to configure client certificate trustpoint (#71)
This patch adjusts the implementation for configuring the grpc server
process on the Cisco NX-OS Device to set a client certificate
trustpoint that is used for mTLS to verify the client certificate
presented by the callers. This corresponds to the cli configuration of:

```
grpc client root certificate <trustpoint-id>
```

Additionally, this patch implements the `Reset` func of the grpc
configuration. It will not disable the grpc server but just reset some
of it's properties to their default value.