Make SecretRef in Device Immutable after creation
The DeviceReconciler will add a finalizer to the referenced secret to
make sure that it's not deleted while still used from the device.
Additionally, it's now also validated that the secret reference, once
set on the device, can not be changed. This prevents the user from
changing the secretRef on a device, leaving the old referenced secret
with the finalizer.