Refactor status conditions and generic error handling
This change makes use of the `conditions` package to universally
update the status conditions of resources. It also introduces a
`FromError` func inside the `utility` package to generically formalize a
`Configured` condition based on an error returned by the provider. This
logic is inspired from the implementation previously done inside the
nx-os provider.