Introduce a periodic requeue mechanism based on a configurable interval
This change introduces a new argument `--requeue-interval` on the
manager program, that defines a duration after which Kubernetes
resources should be reconciled, regardless of whether they have changed.
This duration will be used to set the `RequeueAfter` of the
`ctrl.Result` after a reconciliation, such that state changes can be
detected over time. An example of this is the transition of the
operational state ("oper-status") of an `Interface` that happens over
time. Additionally, this value is used to requeue when waiting for
procedures to finish (e.g. the Device provisioning). A `Jitter` is added
to the value, such that we prevent all resources from being requeued at
the same time.