[NX-OS] Configure interface ip address after the interface entity
This change alters the order of how the individual configuration
entities are applied for a given `Interface` resource inside the Cisco
NX-OS provider. It ensures that the addr entity is configured after the
respective interface entity. This prevents a issue, where configuring a
non-present interface lead to an error on the provider, as the
respective gNMI SetRequest tried to configure the address before the
interface, where the device would return an error.