[NX-OS] Default spanning-tree mode for physical interfaces
This change corrects the deletion logic for the spanning-tree mode on a
physical interface. Instead of trying to delete the YANG tree, which
lead to an error, the value is now reset to a default using the
`gnmiext.Defaultable` interface. Additionally, as the spanning-tree mode
is not concerned with Loopback interfaces, it's not tried to be
finalized there, as it previously lead to an error trying to to
delete/default for a Loopback interface during deletion.