feat: VRF
Adds VRF (Virtual Routing and Forwarding) namespaced CRD to model
per-tenant routing contexts. It follows our brick model with an
immutable ref to a Device. While we add a type to support a
provider-specific config, we don't make any use of it here. We suport
defining a VNI, a RouteDistinguisher, and a list of Route Targets (each
with a value, a list of address families and an action).

We include an admission hook to validate some of the fields as the
fomatting rules for them are not trivial (see Route Distinguishers
encoding in RFC 4364).

The reconciler fails if some features are not enabled in the remote
device and the spec has values that relate to it. For example, the
family type IPv4EVPN requires the `bgp` feature, while the vni field
might require the feature `nv overlay` with the `evpn` setting. Enabling
these features on the device is not responsability of this brick. The
reconciliation will fail and the user must enable the features by other
means. We opt to not update or check if the feature is enabled to reduce
the number of gNMI calls and RTTs.

We update the cisco-nxos-gnmi provider implementation to facilitate
creating the gNMI payload from the CRDs.