Fix e2e test by waiting for Interface to become configured
As far as our E2E tests are concerned, we are only checking if we apply
the correct configuration against our fake gnmi server. As this server
has no notion of how it would handle operational state, we simply wait
for the Kubernetes resource to be configured, ignoring it's operational
state in favor of only verifying if the configuration we applied matches
what we expect.