Remove `provider.Result` Response type from provider integration
This changes removes the `provider.Result` return argument historically
used on the "EnsureX" provider methods. This type contained a
`RequeueAfter` field that mimicked the equally named field of the
`ctrl.Result` returned from the reconciliation loop. The reason for this
changes is that we decided to keep the orchestration flow, incl. the
decision for and how requeuing should be done up to the controller and
should avoid leaking such kubernetes specific logic into the provider
code.