fix: Tiltfile passes an IPv6 addr, which fails validation
Tilt calls `nslookup`, which can return an IPv4 and/or an IPv6
address(es)  with no guarantees about the order. This causes Tilt to
supply an IPv6 address to `DeviceSpec.Endpoint.Address`, which is then
rejected by a kubebuilder validation pattern. This commit fixes this by
querying for IPv4 addresses. The alternative approach of allowing IPv6
to connect to devices is not considered for now.