feat: Virtual Tunnel End-Points (VTEPs)
* Adds generic support for VXLAN Tunnel End Point (VTEP). The core type
  `VTEP` requires at least two loopback interfaces on the same device.
The generic API type includes an optional field for Multicastgroup. The
generic VTEP includes a `ProviderConfigRef` field to support extended
capabilities for specific switches.
* The `VTEP` controller includes Configuration and Operation status
  updates.
* The type `VTEP` has an admission webhook to validate that the provided
  multicast prefix is indeed a multicast prefix. Notice that the current
implementation of this check is partial, and it relies on further check
at the target device.
* This multicast prefix check requires extending the `IPPrefix` type
  with two methods that provide the first and the last IP in the prefix.
* The `cisco.nx.v1alpha1.VTEPConfig` can be used in the
  `ProviderConfigRef` includes cisco specific settings, including system
settings. Thus, updates to a single `VTEPConfig` that is reference by
multiple `VTEP` will result in changing system settings in multiple
devices (infra VLANs).
* The `VTEP` controller checks during reconciliation that the provider
  `ProviderConfigRef` is actually supported. It implements it by
checking the `Device` status fields and looking for the device model
information. It then checks if the API version of the
`ProviderConfigRef` supports this model.
* The `cisco.nx` type `VTEPConfig` has also a webhook to validate some
  of the conditions required by Cisco for the infra VLANs. Even though
the device also checks for these settings, we add the webhook to avoid a
gNMI call to the device when possible.