[NX-OS] Remove implicit copp profile setting
This change removes the implicit configuration of the control plane
policing profile. Currently, we are only interested in having
`copp profile strict` configured. This is already configured by default
when the device boots from a fresh state. Hence, we have no need to
expose this configuration via a custom resource.