Fix `gnmiext` skipping updates due to missing fields in YANG payload
When fields are absent in the YANG payload, they are skipped when the
payload is unmarshaled into a Go struct. This lead to invalid
comparisons, as the `gnmiext.Client` failed to detect changes in this
case. This commit fixes this issue, by resetting struct fields not
present in the YANG payload with a json struct tag back to their zero
value.